IMPROVEMENT-39: Move ssh logic to a separate module
