+++ internal/model/ssh/cmd_win.go
@@ -1,8 +1,8 @@
-//go:build !windows
+//go:build windows
 
-package sshconfig
+package ssh
 
 // BaseCMD return OS specific 'ssh' command.
 func BaseCMD() string {
-	return "ssh"
+	return "cmd /c ssh"
 }
